LayoutTests/storage/indexeddb/keypath-fetch-key-expected.txt

Test IndexedDB keyPaths

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')

testing simple identifier
Deleted all object stores.
objectStore = db.createObjectStore(testData[nextToOpen].name, {keyPath: testData[nextToOpen].keyPath});
result = objectStore.add(testData[nextToOpen].value);
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is testData[nextToOpen].key

testing simple identifiers
Deleted all object stores.
objectStore = db.createObjectStore(testData[nextToOpen].name, {keyPath: testData[nextToOpen].keyPath});
result = objectStore.add(testData[nextToOpen].value);
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is testData[nextToOpen].key

testing nested identifiers
Deleted all object stores.
objectStore = db.createObjectStore(testData[nextToOpen].name, {keyPath: testData[nextToOpen].keyPath});
result = objectStore.add(testData[nextToOpen].value);
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is testData[nextToOpen].key

testing nested identifiers with distractions
Deleted all object stores.
objectStore = db.createObjectStore(testData[nextToOpen].name, {keyPath: testData[nextToOpen].keyPath});
result = objectStore.add(testData[nextToOpen].value);
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is testData[nextToOpen].key
PASS successfullyParsed is true

TEST COMPLETE
